+++ gems/aws-sdk-elasticsearchservice/CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased Changes
 ------------------
 
+* Issue - Replace runtime endpoint resolution approach with generated ruby code.
+
 1.68.0 (2022-11-10)
 ------------------