+++ gems/aws-sdk-managedgrafana/CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased Changes
 ------------------
 
+* Issue - Replace runtime endpoint resolution approach with generated ruby code.
+
 1.10.0 (2022-11-23)
 ------------------