+++ gems/aws-sdk-rekognition/CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased Changes
 ------------------
 
+* Issue - Replace runtime endpoint resolution approach with generated ruby code.
+
 1.73.0 (2022-12-12)
 ------------------